+++ lib/flutter_ilib.dart
@@ -3,6 +3,8 @@ library flutter_ilib;
 import 'package:flutter/foundation.dart';
 import 'package:flutter_js/flutter_js.dart';
 import 'ilib_init.dart';
+import 'internal/logger/log_adapter.dart';
+import 'internal/logger/logger.dart';
 
 export 'ilib_datefmt.dart';
 export 'ilib_init.dart';
@@ -18,6 +20,7 @@ class FlutterILib extends ChangeNotifier {
 
   static final FlutterILib _instance = FlutterILib._internal();
   static FlutterILib get instance => _instance;
+  final LogAdapter logger = Logger();
 
   /// Return whether iLib is ready
   bool get isILibReady => ILibJS.instance.isILibReady;
@@ -43,6 +46,7 @@ class FlutterILib extends ChangeNotifier {
   /// To properly load the updated locale data file,
   /// this should be called at the appropriate time when the locale changes.
   Future<void> loadLocaleData(String? locale) async {
+    logger.debug('[FlutterILib] Loading locale data for $locale');
     await ILibJS.instance.loadILibLocaleData(locale);
   }
 }