+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.0
 * Added `isILibReady` getter to `ILibJS` class to expose internal `_iLibPrepared` state.
 * Added `isILibReady` getter to `FlutterILib` class for external access to ILibJS initialization status.
+* Updated logging system to use the logging package with support for multiple log levels
 
 ## 1.1.0
 * Updated the iLib files to version 14.21.0 since the new version of iLib has been released.